update build process
